updated entities

+++ src/main/java/com/lohith/mypaymentwebapp/repository/TransactionRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.lohith.mypaymentwebapp.entity.TransactionEntity;
 
 @Repository
-public interface TransactionRepository extends JpaRepository <TransactionEntity,Integer>{
+public interface TransactionRepository extends JpaRepository <TransactionEntity,Long>{
 
 }
